+++ packages/siwe/src/client.ts
@@ -50,8 +50,8 @@ export class Web3ModalSIWEClient {
     this.methods = siweConfigMethods
   }
 
-  async getNonce() {
-    const nonce = await this.methods.getNonce()
+  async getNonce(address?: string) {
+    const nonce = await this.methods.getNonce(address)
     if (!nonce) {
       throw new Error('siweControllerClient:getNonce - nonce is undefined')
     }
@@ -85,8 +85,8 @@ export class Web3ModalSIWEClient {
   }
 
   async signIn(): Promise<SIWESession> {
-    const nonce = await this.methods.getNonce()
     const { address } = AccountController.state
+    const nonce = await this.methods.getNonce(address)
     if (!address) {
       throw new Error('An address is required to create a SIWE message.')
     }