+++ apps/gallery/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @apps/gallery
 
+## 4.1.1
+
+### Patch Changes
+
+- Fix siwe version
+
+- Updated dependencies []:
+  - @web3modal/common@4.1.1
+  - @web3modal/ui@4.1.1
+
+## 4.1.0
+
+### Minor Changes
+
+- Email Stable release
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/common@4.1.0
+  - @web3modal/ui@4.1.0
+
 ## 4.0.13
 
 ### Patch Changes